// WEB/Blog/src/app/features/blogposts/models/add-blog-post-request.model.ts
import { Category } from "../../category/models/category.model";

export interface AddBlogPostRequest {
    title: string;
    shortDescription: string;
    content: string;
    featuredImageURL: string;
    urlHandle: string;
    publishedDate: Date;
    author: string;
    isVisible: boolean;
    categories: Category[];
  }
  